Update DIMM memorySizeInKB to be in size_t
Following the changes in
https://gerrit.openbmc.org/c/openbmc/phosphor-dbus-interfaces/+/41870
where `MemorySizeInKB` type changed from `uint32` to `size`.

+++ redfish-core/lib/systems.hpp
@@ -339,7 +339,7 @@ inline void
                                 return;
                             }
 
-                            const uint32_t* memorySizeInKB = nullptr;
+                            const size_t* memorySizeInKB = nullptr;
 
                             const bool success =
                                 sdbusplus::unpackPropertiesNoThrow(
